[backend] add batch create task endpoint
