+++ ymir/backend/src/common/id_definition/error_codes.py
@@ -65,6 +65,7 @@ class APIErrorCode(IntEnum):
     DATASET_NOT_ACCESSIBLE = 110403
     DATASET_FAILED_TO_CREATE = 110404
     DATASET_PROTECTED_TO_DELETE = 110405
+    DATASETS_NOT_IN_SAME_GROUP = 110406
 
     ASSET_NOT_FOUND = 110501
 
@@ -118,3 +119,4 @@ class APIErrorCode(IntEnum):
     FAILED_TO_IMPORT_MODEL = 111801
 
     REFUSE_TO_PROCESS_MIXED_OPERATIONS = 111901
+    FAILED_TO_EVALUATE = 111902
